~:Documents:ADHD_Productivity_App/requirements.txt
streamlit>=1.35.0
anthropic>=0.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0
pandas>=2.0.0
